backend/requirements.txt
asgiref==3.8.1
autopep8==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-extensions==3.2.3
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
pycodestyle==2.11.1
PyJWT==2.8.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
